+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2015-11-01
 - add `format.print_table`
 - use `format.print_table` for `instances` module
+- `instances`: print CF stack name if available
 
 ## 2015-10-14
 - added a license file